requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1


[Unit]
Description=edaily daemon
After=network.target

[Service]
User=soofa
Group=www-data
WorkingDirectory=/home/soofa/d/edaily
ExecStart=/home/soofa/d/edaily/.venv/bin/gunicorn --access-logfile - --workers 3 --bind unix:/home/soofa/d/edaily/gala.sock gala.wsgi:application

[Install]
WantedBy=multi-user.target